feat: Basic*Service.java @Service, @RequiredArgsConstructor 작성

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -5,30 +5,36 @@ import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.MessageRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.MessageService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
 
+@Service("basicMessageService")
+@RequiredArgsConstructor
 public class BasicMessageService implements MessageService {
 
     private final MessageRepository messageRepository;
     private final ChannelRepository channelRepository;
     private final UserRepository userRepository;
 
-    public BasicMessageService(MessageRepository messageRepository, ChannelRepository channelRepository, UserRepository userRepository) {
-        this.messageRepository = messageRepository;
-        this.channelRepository = channelRepository;
-        this.userRepository = userRepository;
-    }
-
     @Override
-    public Message create(String content, UUID userId, UUID channelId) {
-        if (content == null || content.isBlank() || userId == null || channelId == null) {
+    public Message create(String content, UUID channelId, UUID authorId) {
+        if (!channelRepository.existsById(channelId)) {
+            throw new NoSuchElementException("Channel not found with id " + channelId);
+        }
+
+        if (!userRepository.existsById(authorId)) {
+            throw new NoSuchElementException("Author not found with id " + authorId);
+        }
+
+        if (content == null || content.isBlank() || channelId == null || authorId == null) {
             throw new IllegalArgumentException("Message info is invalid");
         }
 
-        Message message = new Message(content, userId, channelId);
+        Message message = new Message(content, channelId, authorId);
         return messageRepository.save(message);
     }
 
@@ -51,6 +57,9 @@ public class BasicMessageService implements MessageService {
 
     @Override
     public boolean delete(UUID messageId) {
+        if (!messageRepository.existsById(messageId)) {
+            throw new NoSuchElementException("Message with id " + messageId + " not found");
+        }
         return messageRepository.delete(messageId);
     }
 }
